Extract push events and persistence features and data
